Container, extractor module, resources on tiles.
[ADDED]     Container module: stocks extracted resources.
[UPDATED]   Distribute extracted resources among connected containers.
[MOVED]     Path finding logic extracted, generalized into PathUtils.
[ADDED]     neighbours utility in PathUtils.
[ADDED]     Resource distribution across the game map.
[ADDED]     Level types determine what resources their tiles support.
[ADDED]     Convenience methods when doing Coordinate stuff on something that knows its level.

+++ omicron-core/src/main/java/com/lyndir/omicron/api/model/Airship.java
@@ -14,7 +14,7 @@ public class Airship extends PlayerObject {
 
         super( "Airship", owningPlayer, locationTile, //
                new BaseModule( 5, 1, 5, ImmutableSet.of( LevelType.SKY ) ), //
-               new MobilityModule( 2, ImmutableMap.of( LevelType.SKY, 1f ), ImmutableMap.<LevelType, Float>of() ) );
+               new MobilityModule( 2, ImmutableMap.of( LevelType.SKY, 1d ), ImmutableMap.<LevelType, Double>of() ) );
     }
 
     @NotNull
